Fixed case of empty json pointer
Currently, we are saying that yacl4j adheres to the Json Pointer specification with one exception: the leading '/' is optional.

This could come in handy for properties-based configurations but has one drawback: "/" and "" are both valid json pointers which map respectively to an empty key at the root and the whole document. Making the leading '/' optional means that we need to add it when it's not specified but we should not add it in case of "".

For now the behavior is consistent between system properties and placeholder resolution but it may change in future, i.e. keeping the "relaxed" Json Pointer specification for system properties and the original one for placeholder resolution.