Refactoring FileConfigurationSource class
The configuration now is handled internally as an InputStream rather than a File.

This also comes in handy when loading a configuration file which is embedded in some other jar on the classpath. Suppose you are developing multiple services which share some code and configuration defaults. Most likely you will end up building a common library for them. The configuration defaults could be placed in a file within the common library and easily loaded as a resource on the classpath.