No more clone of system properties before transformation
System properties were cloned before being transformed into a JsonNode configuration. We are no longer cloning them because there is no real advantage in doing it. Moreover, we have aligned the behavior to the EnvironmentVariablesConfigurationSource class.

It would probably make sense to clone system properties and environment variables in the constructor of the respective sources because that would guarantee consistent results over multiple invocation of the ConfigurationBuilder#build method. But again, that's not always the desired behavior so, in the worst case, we should make this configurable.