Overriding Travis CI build step
It seems that the default build step of Travis CI "mvn test -B" does not work well with the repackaged module.

First, the "shade" goal of the maven-shade-plugin is bound to the "package" phase and cannot be anticipated (according to the error message received after trying). Because the "test" phase happens before the "package" phase, the jar of the repackaged module (i.e. yacl4j-repackaged-jackson) is not available when the compilation of yacl4j-core module happnes. So the "target/classes" folder of yacl4j-repackaged-jackson is used for the compilation of yacl4j-core but that folder is empty because the only purpose of the yacl4j-repackaged-jackson module is to repackage Jackson.

To be decided if this is just a temporary workaround or a definitive fix.